requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
psycopg2-binary==2.9.10
python-anticaptcha==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
2captcha-python==1.1.3